[feat] : 기기 등록 스플래시로 변경
