--- app/src/main/java/com/myongsik/myongsikandroid/util/GetAdvertisingIdTask.kt
@@ -1,26 +1,0 @@
-package com.myongsik.myongsikandroid.util
-
-import android.content.Context
-import android.os.AsyncTask
-import com.google.android.gms.ads.identifier.AdvertisingIdClient
-import com.google.android.gms.common.GooglePlayServicesNotAvailableException
-import com.google.android.gms.common.GooglePlayServicesRepairableException
-import java.io.IOException
-
-class GetAdvertisingIdTask(private val context: Context) :
-    AsyncTask<Void, Void, String>() {
-
-    override fun doInBackground(vararg params: Void?): String? {
-        try {
-            val info = AdvertisingIdClient.getAdvertisingIdInfo(context)
-            return info.id
-        } catch (e: IOException) {
-            e.printStackTrace()
-        } catch (e: GooglePlayServicesNotAvailableException) {
-            e.printStackTrace()
-        } catch (e: GooglePlayServicesRepairableException) {
-            e.printStackTrace()
-        }
-        return null
-    }
-}
