+++ app/src/main/res/values/strings.xml
@@ -6,7 +6,7 @@
     <string name="home_time_life_tv">운영시간 │ 중식 11:30~13:30 | 석식 17:00 ~ 18:30</string>
     <string name="home_time_staff_tv">운영시간 │ 중식 11:30~13:30 | 석식 17:30 ~ 18:30</string>
     <string name="home_time_student_tv">운영시간 │ 조식 8:30~9:30 | 중식 10:00 ~ 15:00</string>
-    <string name="home_time_myonog_tv">운영시간 │ 미정</string>
+    <string name="home_time_myonog_tv">운영시간 │ 백반 11:30~14:30 샐러드,볶음밥 10:00 ~ 15:00</string>
 
     <string name="week_today_tv">주간 식단표</string>
     <string name="splash_help_select">캠퍼스 설정을 통해 \n 접속 시 선택한 캠퍼스 \n 학식 메뉴를 확인할 수 있습니다.</string>
@@ -45,7 +45,7 @@
     <string name="select_student_restaurant">학관 식당</string>
     <string name="select_student_restaurant_time">운영시간\n조식 8:30~9:30 | 중식 10:00 ~ 15:00</string>
     <string name="select_new_restaurant">명진당</string>
-    <string name="select_new_restaurant_time">운영시간\n미정</string>
+    <string name="select_new_restaurant_time">운영시간\n백반 11:30~14:30 \n샐러드,볶음밥 10:00 ~ 15:00</string>
 
     <string name="splash_myongji_title_text">명지대 학식 식단표 제공 앱</string>
     <string name="tag_myongji_good_food">#명지 맛집</string>
